pyproject.toml
[tool.poetry]
name = "devnet"
version = "0.1.0"
description = ""
authors = ["Yuji Kamiya <y.kamiya0@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.1>=,<4.0"
torch = "^1.10.2"
hydra-core = "^1.1.1"
logzero = "^1.7.0"
pandas = "1.3.5"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
